requirements.txt
python-telegram-bot==13.14
python-dotenv~=1.0
environs~=11.0
google-cloud-dialogflow
vk_api~=11.9
